Tweaking main.go for testing
